#74: Moved types to out of the core module
