+++ jodin-common/src/types/mod.rs
@@ -15,16 +15,16 @@ use resolved_type::WeakResolvedType;
 use crate::core::identifier_resolution::{Registrable, Registry};
 use crate::core::privacy::Visibility;
 use crate::core::tags::Tag;
-use crate::core::types::arrays::Array;
-use crate::core::types::jobject::JObject;
-use crate::core::types::pointer::Pointer;
-use crate::core::types::primitives::Primitive;
-use crate::core::types::resolved_type::{ResolveType, ResolvedType};
-use crate::core::types::structure::Structure;
-use crate::core::types::traits::{JTrait, JTraitObject};
-use crate::core::types::type_environment::TypeEnvironment;
 use crate::error::{JodinError, JodinErrorType, JodinResult};
 use crate::identifier::Identifier;
+use crate::types::arrays::Array;
+use crate::types::jobject::JObject;
+use crate::types::pointer::Pointer;
+use crate::types::primitives::Primitive;
+use crate::types::resolved_type::{ResolveType, ResolvedType};
+use crate::types::structure::Structure;
+use crate::types::traits::{JTrait, JTraitObject};
+use crate::types::type_environment::TypeEnvironment;
 use crate::utility::Visitor;
 
 pub mod arrays;
