+++ jodinc/src/passes/analysis/identity_resolution_tool.rs
@@ -10,8 +10,8 @@ use jodin_common::core::import::{Import, ImportType};
 use jodin_common::core::privacy::{Visibility, VisibilityTag};
 use jodin_common::core::tags::TagTools;
 use jodin_common::core::tags::{ResolvedIdentityTag, Tag};
-use jodin_common::core::types::intermediate_type::{IntermediateType, TypeSpecifier, TypeTail};
-use jodin_common::core::types::StorageModifier;
+use jodin_common::types::intermediate_type::{IntermediateType, TypeSpecifier, TypeTail};
+use jodin_common::types::StorageModifier;
 use jodin_common::unit::TranslationUnit;
 use jodin_common::utility::Tree;
 use std::any::Any;
